count/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.23.0
paho-mqtt>=1.6.1
